OMAPDSS: TFP410: Add i2c support
The TFP410 can be controlled by I2C and on some boards it will need to be
turned on by an I2C command before it can be used. This commit adds basic I2C
support to the driver such that it will be switched to normal operation mode
at probe time if it is defined on an I2C bus in the device tree.

Signed-off-by: James Barker <jbarker@elesar.co.uk>