clk: ti: clk-7xx: Allow ABE clock parent to be defined in device tree
The ABE clock parent is hard-coded to be sys_clkin2, but this does not work
on the Elesar Titanium board, where it must be sys_clkin1. This commit adds
a device tree property that can be used to set the clock parent.

This is a temporary workaround as this code has been changed in later
versions of the TI kernel.

Signed-off-by: James Barker <jbarker@elesar.co.uk>