mfd: palmas: Add support for soft power-off on the Elesar Titanium board
The Titanium board uses a GPIO line to trigger power-off. We can't use
the 'gpio-poweroff' driver because we need the Palmas driver to safely
shut the power down and only one driver can use the pm_power_off
function pointer, so instead add functionality to the palmas driver to
trigger the GPIO line when it is shutting the system down.

Signed-off-by: James Byrne <jbyrne@elesar.co.uk>