ARM: dts: DRA7: Allow Etnaviv driver to be selected on DRA7
Extend the 'compatible' type of the 'bb2d' device and add a
'gpu-subsystem' entry to the DRA7 device tree so that the Etnaviv driver
can be used with these devices.

Also remove an invalid dependency from the Etnaviv Kconfig file so that
the driver can be selected on Arm devices.

Signed-off-by: James Byrne <jbyrne@elesar.co.uk>